Abstract server sent events logic to its own file
